Assignment1_Problem3
The objective is to write a program that determines the nearest point to a test point in a vector of points. It is required to create an interface and then implement that interface. The program must also give the option to calculate the distance between points using the Euclidean or Manhattan Distances.